COM-2642: reorganize utils
