+++ src/core/components/courses/SlotContainer.vue
@@ -72,8 +72,8 @@ import SlotEditionModal from '@components/courses/SlotEditionModal';
 import SlotCreationModal from '@components/courses/SlotCreationModal';
 import { NotifyNegative, NotifyWarning, NotifyPositive } from '@components/popup/notify';
 import { E_LEARNING, ON_SITE, REMOTE } from '@data/constants';
-import { formatQuantity, formatDuration, formatIntervalHourly, getDuration } from '@helpers/utils';
-import { formatDate } from '@helpers/date';
+import { formatQuantity } from '@helpers/utils';
+import { formatDate, formatDuration, formatIntervalHourly, getDuration } from '@helpers/date';
 import { frAddress, minDate, maxDate, urlAddress } from '@helpers/vuelidateCustomVal';
 import moment from '@helpers/moment';
 import { courseMixin } from '@mixins/courseMixin';
