+++ src/modules/client/mixins/payMixin.js
@@ -3,8 +3,8 @@ import omit from 'lodash/omit';
 import Pay from '@api/Pay';
 import { NotifyPositive, NotifyNegative } from '@components/popup/notify';
 import { END_CONTRACT_REASONS, SURCHARGES } from '@data/constants';
-import { formatPrice, formatIdentity, formatHours, formatNumberForCSV } from '@helpers/utils';
-import { ascendingSort } from '@helpers/date';
+import { formatPrice, formatIdentity, formatNumberForCSV } from '@helpers/utils';
+import { ascendingSort, formatHours } from '@helpers/date';
 import { downloadCsv, downloadFile } from '@helpers/file';
 import moment from '@helpers/moment';
