+++ src/modules/client/pages/ni/auxiliaries/Dashboard.vue
@@ -126,7 +126,8 @@ import Events from '@api/Events';
 import Stats from '@api/Stats';
 import { NotifyNegative } from '@components/popup/notify';
 import { AUXILIARY_ROLES, DEFAULT_AVATAR } from '@data/constants';
-import { formatHours, roundFrenchPercentage } from '@helpers/utils';
+import { roundFrenchPercentage } from '@helpers/utils';
+import { formatHours } from '@helpers/date';
 import moment from '@helpers/moment';
 import Gauge from 'src/modules/client/components/Gauge';
 import ChipsAutocomplete from 'src/modules/client/components/planning/ChipsAutocomplete';
