+++ src/modules/client/pages/ni/pay/Absences.vue
@@ -48,9 +48,9 @@ import Button from '@components/Button';
 import DateRange from '@components/form/DateRange';
 import TitleHeader from '@components/TitleHeader';
 import SimpleTable from '@components/table/SimpleTable';
-import { formatIdentity, formatHours, formatHoursWithMinutes } from '@helpers/utils';
+import { formatIdentity } from '@helpers/utils';
 import moment from '@helpers/moment';
-import { formatDate } from '@helpers/date';
+import { formatDate, formatHours, formatHoursWithMinutes } from '@helpers/date';
 import { ABSENCE, ABSENCE_NATURES, ABSENCE_TYPES, DAILY, AUXILIARY } from '@data/constants';
 import EventEditionModal from 'src/modules/client/components/planning/EventEditionModal';
 import { planningActionMixin } from 'src/modules/client/mixins/planningActionMixin';
